app_belvo/requirements.txt
fastapi
firebase-admin
pyrebase4
uvicorn
pydantic
pydantic[email]
fastapi[all]